+++ Source/EventFlow/Aggregates/IDomainEvent.cs
@@ -29,7 +29,6 @@ namespace EventFlow.Aggregates
         Type AggregateType { get; }
         Type EventType { get; }
         int AggregateSequenceNumber { get; }
-        Guid BatchId { get; }
         IMetadata Metadata { get; }
         DateTimeOffset Timestamp { get; }