+++ Source/EventFlow/EventStores/EventJsonSerializer.cs
@@ -60,7 +60,8 @@ namespace EventFlow.EventStores
             return new SerializedEvent(
                 metaJson,
                 dataJson,
-                metadata.AggregateSequenceNumber);
+                metadata.AggregateSequenceNumber,
+                metadata);
         }
 
         public IDomainEvent Deserialize(ICommittedDomainEvent committedDomainEvent)
@@ -77,8 +78,7 @@ namespace EventFlow.EventStores
                 aggregateEvent,
                 metadata,
                 committedDomainEvent.AggregateId,
-                committedDomainEvent.AggregateSequenceNumber,
-                committedDomainEvent.BatchId);
+                committedDomainEvent.AggregateSequenceNumber);
 
             return domainEvent;
         }