+++ RELEASE_NOTES.md
@@ -18,6 +18,8 @@
      a `GlobalSequenceNumberRange` removed
  * Breaking: Remove the concept of event caches. If you really need this
    then implement it by registering a decorator for `IEventStore`
+ * Breaking: Moved `IDomainEvent.BatchId` to metadata and created
+   `MetadataKeys.BatchId` to help access it
  * New: `IEventStore.DeleteAggregateAsync` to delete an entire aggregate
    stream. Please consider carefully if you really want to use it. Storage
    might be cheaper than the historic knowledge within your events